+++ src/lib/common/component_utilities.js
@@ -45,14 +45,32 @@ export function greatestVersionDesc(t1, t2) {
  * @param {import('$lib/types').Task} t2
  * @returns {-1|0|1}
  */
-function compareTaskNameAndVersion(t1, t2) {
+export function compareTaskNameAscAndVersionAsc(t1, t2) {
 	if (t1.name < t2.name) return -1;
 	if (t1.name > t2.name) return 1;
 	// Names are equal, sort by version
 	return greatestVersionAsc(t1, t2);
 }
 
-export function orderTasksByOwnerThenByNameThenByVersion(tasks, ownerName = null) {
+/**
+ * @param {import('$lib/types').Task} t1
+ * @param {import('$lib/types').Task} t2
+ * @returns {-1|0|1}
+ */
+export function compareTaskNameAscAndVersionDesc(t1, t2) {
+	if (t1.name < t2.name) return -1;
+	if (t1.name > t2.name) return 1;
+	// Names are equal, sort by version
+	return greatestVersionDesc(t1, t2);
+}
+
+/**
+ * @param {import('$lib/types').Task[]} tasks
+ * @param {string|null} ownerName
+ * @param {'asc'|'desc'} order
+ */
+export function orderTasksByOwnerThenByNameThenByVersion(tasks, ownerName = null, order = 'asc') {
+	const sortingFunction = getVersionSortingFunction(order);
 	// Sort tasks by owner, by name and by version
 	return tasks.sort((t1, t2) => {
 		// If ownerName is not null, filter tasks by owner
@@ -61,7 +79,7 @@ export function orderTasksByOwnerThenByNameThenByVersion(tasks, ownerName = null
 			if (t1.owner === ownerName) {
 				if (t2.owner !== ownerName) return -1;
 				// Both owners are same, sort by name and version
-				return compareTaskNameAndVersion(t1, t2);
+				return sortingFunction(t1, t2);
 			} else {
 				// t1 owner is not same as ownerName, t2 should go before t1
 				if (t2.owner === ownerName) return 1;
@@ -72,7 +90,7 @@ export function orderTasksByOwnerThenByNameThenByVersion(tasks, ownerName = null
 			// Should check if t2 owner is null too
 			if (t2.owner === null) {
 				// Both owners are null, sort by name
-				return compareTaskNameAndVersion(t1, t2);
+				return sortingFunction(t1, t2);
 			} else {
 				// t1 owner is null, t2 owner is not null, t1 should go before t2
 				return -1;
@@ -84,9 +102,19 @@ export function orderTasksByOwnerThenByNameThenByVersion(tasks, ownerName = null
 			if (t1.owner < t2.owner) return -1;
 			if (t1.owner > t2.owner) return 1;
 			// Owners are equal, sort by name
-			return compareTaskNameAndVersion(t1, t2);
+			return sortingFunction(t1, t2);
 		}
 	});
+}
+
+/**
+ * @param {'asc'|'desc'} order
+ */
+function getVersionSortingFunction(order) {
+	if (order === 'asc') {
+		return compareTaskNameAscAndVersionAsc;
+	}
+	return compareTaskNameAscAndVersionDesc;
 }
 
 export function fieldHasValue(event) {