+++ src/lib/components/tasks/TaskCollection.svelte
@@ -1,6 +1,6 @@
 <script>
-	import { onMount } from 'svelte';
-	import { browser } from '$app/environment';
+	import { env } from '$env/dynamic/public';
+	import { onDestroy, onMount } from 'svelte';
 	import { collectTaskErrorStore } from '$lib/stores/errorStores';
 	import { modalTaskCollectionId } from '$lib/stores/taskStores';
 	import TaskCollectionLogsModal from '$lib/components/tasks/TaskCollectionLogsModal.svelte';
@@ -26,21 +26,40 @@
 	// If a collection status is installing, the component shall fetch update
 
 	// Component properties
+	/** @type {import('$lib/types').TasksCollections[]} */
 	let taskCollections = [];
+	/** @type {import('$lib/types').TasksCollectionsStateData|undefined} */
 	let taskCollectionAlreadyPresent = undefined;
+
+	/** @type {'pypi'|'local'} */
+	export let packageType = 'pypi';
+
+	/** @type {() => Promise<void>} */
+	export let reloadTaskList;
 
 	let python_package = '';
 	let package_version = '';
 	let python_version = '';
 	let package_extras = '';
+	/** @type {{key: string, value: string}[]} */
+	let pinnedPackageVersions = [];
+
+	const updateTasksCollectionInterval = env.PUBLIC_UPDATE_JOBS_INTERVAL
+		? parseInt(env.PUBLIC_UPDATE_JOBS_INTERVAL)
+		: 3000;
+	let updateTasksCollectionTimeout = undefined;
 
 	// On component load set the taskCollections from the local storage
 	onMount(async () => {
-		if (browser) {
-			taskCollections = loadTaskCollectionsFromStorage();
-		}
+		taskCollections = loadTaskCollectionsFromStorage();
 		await updateTaskCollectionsState();
+		updateTasksCollectionTimeout = setTimeout(
+			updateTasksCollectionInBackground,
+			updateTasksCollectionInterval
+		);
 	});
+
+	let taskCollectionInProgress = false;
 
 	/**
 	 * Requests a task collection to the server
@@ -60,15 +79,24 @@
 			package_extras
 		};
 
+		const ppv = getPinnedPackageVersionsMap();
+		if (ppv) {
+			requestData.pinned_package_versions = ppv;
+		}
+
+		taskCollectionInProgress = true;
 		const response = await fetch('/api/v1/task/collect/pip', {
 			method: 'POST',
 			credentials: 'include',
 			headers: headers,
 			body: JSON.stringify(requestData, replaceEmptyStrings)
 		});
+		taskCollectionInProgress = false;
 
-		const result = await response.json();
 		if (response.ok) {
+			const result = /** @type {import('$lib/types').TasksCollectionsState} */ (
+				await response.json()
+			);
 			if (response.status === 200) {
 				console.log('Task collection already exists');
 				taskCollectionAlreadyPresent = result.data;
@@ -88,12 +116,34 @@
 			package_version = '';
 			python_version = '';
 			package_extras = '';
+			pinnedPackageVersions = [];
 		} else {
+			const result = await response.json();
 			console.error('Task collection request failed: ', result);
 			collectTaskErrorStore.set(result);
 		}
 	}
 
+	/**
+	 * @returns {{[key: string]: string}|undefined}
+	 */
+	function getPinnedPackageVersionsMap() {
+		/** @type {{[key: string]: string}} */
+		const map = {};
+		for (const ppv of pinnedPackageVersions) {
+			if (ppv.key && ppv.value) {
+				map[ppv.key] = ppv.value;
+			}
+		}
+		if (Object.keys(map).length === 0) {
+			return undefined;
+		}
+		return map;
+	}
+
+	/**
+	 * @param {import('$lib/types').TasksCollectionsState} taskCollection
+	 */
 	function storeCreatedTaskCollection(taskCollection) {
 		taskCollections.push({
 			id: taskCollection.id,
@@ -107,8 +157,8 @@
 
 	/**
 	 * Fetches a task collection from the server
-	 * @param {string} taskCollectionId
-	 * @returns {Promise<*>}
+	 * @param {number} taskCollectionId
+	 * @returns {Promise<import('$lib/types').TasksCollectionsState|undefined>}
 	 */
 	async function getTaskCollection(taskCollectionId) {
 		const response = await fetch(`/api/v1/task/collect/${taskCollectionId}?verbose=True`, {
@@ -133,62 +183,81 @@
 		}
 	}
 
-	async function updateTaskCollectionsState() {
-		const updates = await Promise.allSettled(taskCollections.map((tc) => getTaskCollection(tc.id)));
+	/**
+	 * @param {import('$lib/types').TasksCollections[]|null} collectionsToUpdate
+	 */
+	async function updateTaskCollectionsState(collectionsToUpdate = null) {
+		const collections = collectionsToUpdate ?? taskCollections;
+		const updates = await Promise.allSettled(collections.map((tc) => getTaskCollection(tc.id)));
+
+		const failure = /** @type {PromiseRejectedResult|undefined} */ (
+			updates.find((u) => u.status === 'rejected')
+		);
+		if (failure) {
+			collectTaskErrorStore.set(failure.reason);
+		}
+
+		const successfulUpdates =
+			/** @type {PromiseFulfilledResult<import('$lib/types').TasksCollectionsState|undefined>[]} */ (
+				updates.filter((u) => u.status === 'fulfilled')
+			).map((u) => u.value);
 
 		const updatedTaskCollections = [];
-		for (const update of updates) {
-			if (update.status === 'rejected') {
-				collectTaskErrorStore.set(update.reason);
-			} else {
-				const updatedTaskCollection = update.value;
-				if (!updatedTaskCollection) {
-					// Removing missing task collection
-					continue;
-				}
-				const oldTaskCollection = taskCollections.find((tc) => tc.id === updatedTaskCollection.id);
-				if (!oldTaskCollection) {
-					continue;
-				}
-				oldTaskCollection.status = updatedTaskCollection.data.status;
-				oldTaskCollection.logs = updatedTaskCollection.data.logs;
+
+		for (const oldTaskCollection of taskCollections) {
+			const updatedTaskCollection = successfulUpdates.find(
+				(u) => u !== undefined && u.id === oldTaskCollection.id
+			);
+			if (!updatedTaskCollection) {
 				updatedTaskCollections.push(oldTaskCollection);
+				continue;
 			}
+			oldTaskCollection.status = updatedTaskCollection.data.status;
+			oldTaskCollection.logs = updatedTaskCollection.data.logs;
+			updatedTaskCollections.push(oldTaskCollection);
 		}
 
 		// Update task collections list
 		updateTaskCollections(updatedTaskCollections);
 	}
 
+	/**
+	 * @returns {import('$lib/types').TasksCollections[]}
+	 */
 	function loadTaskCollectionsFromStorage() {
-		if (browser) {
-			// Parse local storage task collections value
-			// @ts-ignore
-			return JSON.parse(window.localStorage.getItem(LOCAL_STORAGE_TASK_COLLECTIONS)) || [];
+		// Parse local storage task collections value
+		const storageContent = window.localStorage.getItem(LOCAL_STORAGE_TASK_COLLECTIONS);
+		if (storageContent) {
+			return JSON.parse(storageContent) || [];
 		}
-		// Fallback to empty task collections list
 		return [];
 	}
 
+	/**
+	 * @param {import('$lib/types').TasksCollections[]} updatedCollectionTasks
+	 */
 	function updateTaskCollections(updatedCollectionTasks) {
-		if (browser) {
-			window.localStorage.setItem(
-				LOCAL_STORAGE_TASK_COLLECTIONS,
-				JSON.stringify(updatedCollectionTasks)
-			);
-			taskCollections = updatedCollectionTasks;
-		}
+		window.localStorage.setItem(
+			LOCAL_STORAGE_TASK_COLLECTIONS,
+			JSON.stringify(updatedCollectionTasks)
+		);
+		taskCollections = updatedCollectionTasks;
 	}
 
 	async function clearTaskCollections() {
 		updateTaskCollections([]);
 	}
 
+	/**
+	 * @param {number} taskCollectionId
+	 */
 	function removeTaskCollection(taskCollectionId) {
 		updateTaskCollections(taskCollections.filter((tc) => tc.id != taskCollectionId));
 	}
 
-	// Component utilities
+	/**
+	 * @param {import('$lib/types').TaskCollectStatus} status
+	 */
 	function statusBadge(status) {
 		switch (status.toLowerCase()) {
 			case 'pending':
@@ -207,6 +276,42 @@
 		const id = event.currentTarget.getAttribute('data-fc-tc');
 		modalTaskCollectionId.set(id);
 	}
+
+	function addPackageVersion() {
+		pinnedPackageVersions = [...pinnedPackageVersions, { key: '', value: '' }];
+	}
+
+	/**
+	 * @param {number} index
+	 */
+	function removePackageVersion(index) {
+		pinnedPackageVersions = pinnedPackageVersions.filter((_, i) => i !== index);
+	}
+
+	async function updateTasksCollectionInBackground() {
+		const collectionsToCheck = taskCollections.filter(
+			(t) => t.status !== 'OK' && t.status !== 'fail'
+		);
+		if (collectionsToCheck.length > 0) {
+			const collectionsToCheckIds = collectionsToCheck.map((c) => c.id);
+			await updateTaskCollectionsState(collectionsToCheck);
+			const newOkTasks = taskCollections.filter(
+				(t) => collectionsToCheckIds.includes(t.id) && t.status === 'OK'
+			).length;
+			if (newOkTasks > 0) {
+				await reloadTaskList();
+			}
+		}
+		clearTimeout(updateTasksCollectionTimeout);
+		updateTasksCollectionTimeout = setTimeout(
+			updateTasksCollectionInBackground,
+			updateTasksCollectionInterval
+		);
+	}
+
+	onDestroy(() => {
+		clearTimeout(updateTasksCollectionTimeout);
+	});
 </script>
 
 <TaskCollectionLogsModal />
@@ -219,68 +324,148 @@
 		</div>
 	{/if}
 	<form on:submit|preventDefault={handleTaskCollection}>
-		<div class="row g-3">
-			<div class="col-6">
+		<div class="row">
+			<div
+				class="mb-2"
+				class:col-md-6={packageType === 'pypi'}
+				class:col-md-12={packageType === 'local'}
+			>
 				<div class="input-group">
 					<div class="input-group-text">
-						<span class="font-monospace">Package</span>
+						<label class="font-monospace" for="package">Package</label>
 					</div>
 					<input
 						name="package"
+						id="package"
 						type="text"
 						class="form-control"
 						required
 						bind:value={python_package}
 					/>
 				</div>
-			</div>
-			<div class="col-6">
-				<div class="input-group">
-					<div class="input-group-text">
-						<span class="font-monospace">Package Version</span>
-					</div>
-					<input
-						name="package_version"
-						type="text"
-						class="form-control"
-						bind:value={package_version}
-					/>
+				<div class="form-text">
+					{#if packageType === 'pypi'}
+						The name of a package published on PyPI
+					{:else}
+						The full path to a wheel file
+					{/if}
 				</div>
 			</div>
-			<div class="col-6">
+			{#if packageType === 'pypi'}
+				<div class="col-md-6 mb-2">
+					<div class="input-group">
+						<div class="input-group-text">
+							<label class="font-monospace" for="package_version">Package Version</label>
+						</div>
+						<input
+							id="package_version"
+							name="package_version"
+							type="text"
+							class="form-control"
+							bind:value={package_version}
+						/>
+					</div>
+				</div>
+			{/if}
+		</div>
+		<div class="row mb-2 mt-2">
+			<div class="col">
+				<span class="fw-bold text-secondary">Optional arguments</span>
+			</div>
+		</div>
+		<div class="row">
+			<div class="col-md-6 mb-2">
 				<div class="input-group">
 					<div class="input-group-text">
-						<span class="font-monospace">Python Version</span>
+						<label class="font-monospace" for="python_version">Python Version</label>
 					</div>
 					<input
+						id="python_version"
 						name="python_version"
 						type="text"
 						class="form-control"
 						bind:value={python_version}
 					/>
 				</div>
+				<div class="form-text">Python version to install and run the package tasks</div>
 			</div>
-			<div class="col-6">
+			<div class="col-md-6 mb-2">
 				<div class="input-group">
 					<div class="input-group-text">
-						<span class="font-monospace">Package extras</span>
+						<label class="font-monospace" for="package_extras">Package extras</label>
 					</div>
 					<input
+						id="package_extras"
 						name="package_extras"
 						type="text"
 						class="form-control"
 						bind:value={package_extras}
 					/>
 				</div>
+				<div class="form-text">
+					Package extras to include in the <code>pip install</code> command
+				</div>
 			</div>
+		</div>
+		{#if pinnedPackageVersions.length > 0}
+			<p class="mt-2">Pinned packages versions:</p>
+		{/if}
+		{#each pinnedPackageVersions as ppv, i}
+			<div class="row">
+				<div class="col-xl-6 col-lg-8 col-md-12 mb-2">
+					<div class="input-group">
+						<label class="input-group-text" for="ppv_key_{i}">Name</label>
+						<input
+							type="text"
+							class="form-control"
+							id="ppv_key_{i}"
+							bind:value={ppv.key}
+							required
+						/>
+						<label class="input-group-text" for="ppv_value_{i}">Version</label>
+						<input
+							type="text"
+							class="form-control"
+							id="ppv_value_{i}"
+							bind:value={ppv.value}
+							required
+						/>
+						<button
+							class="btn btn-outline-secondary"
+							type="button"
+							id="ppv_remove_{i}"
+							aria-label="Remove pinned package version"
+							on:click|preventDefault={() => removePackageVersion(i)}
+						>
+							<i class="bi bi-trash" />
+						</button>
+					</div>
+				</div>
+			</div>
+		{/each}
+		<div class="row">
+			<div class="col-12 mb-1">
+				<button class="btn btn-light" on:click|preventDefault={addPackageVersion}>
+					<i class="bi bi-plus-circle" /> Add pinned package version
+				</button>
+			</div>
+		</div>
+
+		<div class="row">
 			<div class="col-auto">
-				<button type="submit" class="btn btn-primary">Collect</button>
+				<button type="submit" class="btn btn-primary mt-3 mb-3" disabled={taskCollectionInProgress}>
+					{#if taskCollectionInProgress}
+						<div class="spinner-border spinner-border-sm" role="status">
+							<span class="visually-hidden">Collecting...</span>
+						</div>
+					{/if}
+					Collect
+				</button>
 			</div>
 		</div>
 	</form>
 	{#if taskCollections.length > 0}
-		<hr />
-		<div class="">
+		<div class="mb-5">
 			<table class="table table-hover caption-top align-middle">
 				<caption class="text-bg-light border-top border-bottom pe-3 ps-3">
 					<div class="d-flex align-items-center justify-content-between">
@@ -294,9 +479,6 @@
 								message="Clear task collections requests"
 								callbackAction={clearTaskCollections}
 							/>
-							<button class="btn btn-primary" on:click={updateTaskCollectionsState}>
-								Refresh <i class="bi bi-arrow-clockwise" />
-							</button>
 						</div>
 					</div>
 				</caption>