+++ src/routes/tasks/+page.svelte
@@ -1,9 +1,6 @@
 <script>
 	import { page } from '$app/stores';
-	import {
-		orderTasksByOwnerThenByNameThenByVersion,
-		replaceEmptyStrings
-	} from '$lib/common/component_utilities.js';
+	import { orderTasksByOwnerThenByNameThenByVersion } from '$lib/common/component_utilities.js';
 	import { collectTaskErrorStore } from '$lib/stores/errorStores';
 	import { originalTaskStore, taskStore } from '$lib/stores/taskStores';
 	import TaskEditModal from '$lib/components/tasks/TaskEditModal.svelte';
@@ -11,91 +8,79 @@
 	import TaskCollection from '$lib/components/tasks/TaskCollection.svelte';
 	import ConfirmActionButton from '$lib/components/common/ConfirmActionButton.svelte';
 	import { AlertError, displayStandardErrorAlert } from '$lib/common/errors';
+	import AddSingleTask from '$lib/components/tasks/AddSingleTask.svelte';
 
 	// Error property to be set in order to show errors in UI
 	let errorReasons = undefined;
-	// Tasks property updated with respect to data store
-	let tasks = $page.data.tasks;
-	let taskCreateSuccess = false;
 
-	// Add a single task fields
-	let name = '';
-	let command = '';
-	let source = '';
-	let version = '';
-	let input_type = '';
-	let output_type = '';
+	// Tasks property updated with respect to data store
+	/** @type {import('$lib/types').Task[]} */
+	let tasks = $page.data.tasks;
+
+	/** @type {'pypi'|'local'|'single'} */
+	let packageType = 'pypi';
 
 	// Store subscriptions
 	collectTaskErrorStore.subscribe((error) => {
 		if (error) setErrorReasons(error);
 	});
 
-	// Sort tasks
-	tasks = orderTasksByOwnerThenByNameThenByVersion(tasks);
-
 	function setErrorReasons(value) {
 		errorReasons = value;
 		displayStandardErrorAlert(errorReasons, 'errorSection');
 	}
 
-	function setTaskModal(event) {
-		const taskId = event.currentTarget.getAttribute('data-fc-task');
-		const task = tasks.find((t) => t.id == taskId);
+	/**
+	 * @param {import('$lib/types').Task[]} tasks
+	 */
+	function sortTasks(tasks) {
+		return orderTasksByOwnerThenByNameThenByVersion(tasks, null, 'desc');
+	}
+
+	/**
+	 * @param {number} taskId
+	 */
+	function setTaskModal(taskId) {
+		const task = /** @type {import('$lib/types').Task} */ (tasks.find((t) => t.id === taskId));
 		taskStore.set(task);
 		originalTaskStore.set({ ...task });
 	}
 
-	// Updates the tasks list after a task is edited in the modal
+	/**
+	 * Updates the tasks list after a task is edited in the modal
+	 * @param {import('$lib/types').Task} editedTask
+	 */
 	async function updateEditedTask(editedTask) {
-		tasks = tasks.filter((t) => {
+		const updatedTasks = tasks.filter((t) => {
 			if (t.id === editedTask.id) {
 				return editedTask;
 			} else {
 				return t;
 			}
 		});
-	}
-
-	async function reloadTaskList() {
-		window.location.reload();
+		sortTasks(updatedTasks);
+		tasks = updatedTasks;
 	}
 
 	/**
-	 * Creates a new task in the server
-	 * @returns {Promise<*>}
+	 * @param {import('$lib/types').Task} task
 	 */
-	async function handleCreateTask() {
-		taskCreateSuccess = false;
+	function addNewTask(task) {
+		const updatedTasks = [...tasks, task];
+		sortTasks(updatedTasks);
+		tasks = updatedTasks;
+	}
 
-		const headers = new Headers();
-		headers.append('Content-Type', 'application/json');
-
-		const response = await fetch('/api/v1/task', {
-			method: 'POST',
-			credentials: 'include',
-			headers,
-			body: JSON.stringify(
-				{
-					name,
-					command,
-					version,
-					source,
-					input_type,
-					output_type
-				},
-				replaceEmptyStrings
-			)
+	async function reloadTaskList() {
+		const response = await fetch(`/api/v1/task`, {
+			method: 'GET',
+			credentials: 'include'
 		});
-
 		const result = await response.json();
 		if (response.ok) {
-			// Add created task to the list
-			console.log('Task created', result);
-			tasks = [...tasks, result];
-			taskCreateSuccess = true;
+			sortTasks(result);
+			tasks = result;
 		} else {
-			console.error('Unable to create task', result);
 			setErrorReasons(result);
 		}
 	}
@@ -119,124 +104,148 @@
 			throw new AlertError(result);
 		}
 	}
+
+	/**
+	 * @param {number} index
+	 */
+	function isOldVersion(index) {
+		if (index === 0) {
+			return false;
+		}
+		const currentTask = tasks[index];
+		const previousTask = tasks[index - 1];
+		return previousTask.name === currentTask.name && previousTask.owner === currentTask.owner;
+	}
+
+	/**
+	 * @param {number} index
+	 */
+	function isLastOldVersion(index) {
+		if (!isOldVersion(index)) {
+			return false;
+		}
+		if (index === tasks.length - 1) {
+			return false;
+		}
+		const currentTask = tasks[index];
+		const nextTask = tasks[index + 1];
+		return nextTask.name !== currentTask.name || nextTask.owner !== currentTask.owner;
+	}
+
+	/**
+	 * @param {number} index
+	 */
+	function isMainVersion(index) {
+		if (isOldVersion(index)) {
+			return false;
+		}
+		if (index === tasks.length - 1) {
+			return false;
+		}
+		const currentTask = tasks[index];
+		const nextTask = tasks[index + 1];
+		return nextTask.name === currentTask.name && nextTask.owner === currentTask.owner;
+	}
+
+	/**
+	 * @param {Event} event
+	 */
+	function handleToggleOldVersions(event) {
+		const element = /** @type {HTMLElement} */ (event.target);
+		/** @type {HTMLElement|null} */
+		let row = /** @type {HTMLElement} */ (element.closest('tr'));
+		if (!row.classList.contains('expanded')) {
+			closeAllOldVersions(/** @type {HTMLElement} */ (row.closest('table')));
+		}
+		toggleOldVersions(row);
+	}
+
+	/**
+	 * @param {HTMLElement} table
+	 */
+	function closeAllOldVersions(table) {
+		const rows = table.querySelectorAll('tr');
+		for (const row of rows) {
+			if (row.classList.contains('expanded')) {
+				toggleOldVersions(row);
+			}
+		}
+	}
+
+	/**
+	 * @param {HTMLElement} mainRow
+	 */
+	function toggleOldVersions(mainRow) {
+		mainRow.classList.toggle('expanded');
+		/** @type {HTMLElement|null} */
+		let row = mainRow;
+		while ((row = /** @type {HTMLElement|null} */ (row?.nextSibling))) {
+			if (row.classList.contains('old-version')) {
+				row.classList.toggle('collapsed');
+			} else {
+				break;
+			}
+		}
+	}
 </script>
 
-<div class="d-flex justify-content-between align-items-center">
-	<nav aria-label="breadcrumb">
-		<ol class="breadcrumb">
-			<li class="breadcrumb-item active" aria-current="page">Tasks</li>
-		</ol>
-	</nav>
-</div>
-
 <div class="container">
-	<div class="mb-3" id="errorSection" />
+	<div class="d-flex justify-content-between align-items-center">
+		<nav aria-label="breadcrumb">
+			<ol class="breadcrumb">
+				<li class="breadcrumb-item active" aria-current="page">Tasks</li>
+			</ol>
+		</nav>
+	</div>
+
+	<div class="mb-2" id="errorSection" />
+
 	<p class="lead">Add tasks</p>
-	<div class="accordion">
-		<div class="accordion-item">
-			<h2 class="accordion-header">
-				<button
-					class="accordion-button collapsed"
-					data-bs-toggle="collapse"
-					data-bs-target="#taskCollection"
-				>
-					Collect tasks from a package
-				</button>
-			</h2>
-			<div id="taskCollection" class="accordion-collapse collapse">
-				<div class="accordion-body">
-					<TaskCollection />
-				</div>
-			</div>
-		</div>
-		<div class="accordion-item">
-			<h2 class="accordion-header">
-				<button
-					class="accordion-button collapsed"
-					data-bs-toggle="collapse"
-					data-bs-target="#addTask"
-				>
-					Add a single task
-				</button>
-			</h2>
-			<div id="addTask" class="accordion-collapse collapse">
-				<div class="accordion-body">
-					{#if taskCreateSuccess}
-						<div class="alert alert-success" role="alert">Task created successfully</div>
-					{/if}
-					<form on:submit|preventDefault={handleCreateTask}>
-						<div class="row g-3">
-							<div class="col-6">
-								<div class="input-group">
-									<div class="input-group-text">Task name</div>
-									<input name="name" type="text" class="form-control" bind:value={name} />
-								</div>
-							</div>
-							<div class="col-12">
-								<div class="input-group">
-									<div class="input-group-text">Command</div>
-									<input name="command" type="text" class="form-control" bind:value={command} />
-								</div>
-							</div>
-							<div class="col-6">
-								<div class="input-group">
-									<div class="input-group-text">Source</div>
-									<input name="source" type="text" class="form-control" bind:value={source} />
-								</div>
-							</div>
-							<div class="col-6">
-								<div class="input-group">
-									<div class="input-group-text">Version</div>
-									<input name="version" type="text" class="form-control" bind:value={version} />
-								</div>
-							</div>
-							<div class="row" />
-							<div class="col-6">
-								<div class="input-group">
-									<div class="input-group-text">Input type</div>
-									<input
-										name="input_type"
-										type="text"
-										class="form-control"
-										bind:value={input_type}
-									/>
-								</div>
-							</div>
-							<div class="col-6">
-								<div class="input-group">
-									<div class="input-group-text">Output type</div>
-									<input
-										name="output_type"
-										type="text"
-										class="form-control"
-										bind:value={output_type}
-									/>
-								</div>
-							</div>
-							<div class="col-auto">
-								<button type="submit" class="btn btn-primary">Create</button>
-							</div>
-						</div>
-					</form>
-				</div>
-			</div>
-		</div>
+
+	<div class="form-check-inline">Package type:</div>
+
+	<input
+		class="btn-check"
+		type="radio"
+		name="pypi"
+		id="pypi"
+		value="pypi"
+		bind:group={packageType}
+	/>
+	<label class="btn btn-outline-secondary" for="pypi"> PyPI </label>
+
+	<input
+		class="btn-check"
+		type="radio"
+		name="local"
+		id="local"
+		value="local"
+		bind:group={packageType}
+	/>
+	<label class="btn btn-outline-secondary" for="local"> Local </label>
+
+	<input
+		class="btn-check"
+		type="radio"
+		name="single"
+		id="single"
+		value="single"
+		bind:group={packageType}
+	/>
+	<label class="btn btn-outline-secondary" for="single"> Single task </label>
+
+	<div class="mt-3">
+		{#if packageType === 'pypi' || packageType === 'local'}
+			<TaskCollection {packageType} {reloadTaskList} />
+		{:else}
+			<AddSingleTask {addNewTask} />
+		{/if}
 	</div>
 
 	<div class="row mt-4">
 		<p class="lead">Task List</p>
 		<div class="col-12">
-			<table class="table caption-top align-middle">
-				<caption class="text-bg-light border-top border-bottom pe-3 ps-3">
-					<div class="d-flex align-items-center justify-content-between">
-						<span class="fw-normal" />
-						<div>
-							<button class="btn btn-outline-primary" on:click={reloadTaskList}>
-								<i class="bi bi-arrow-clockwise" />
-							</button>
-						</div>
-					</div>
-				</caption>
+			<table class="table align-middle">
 				<thead class="table-light">
 					<tr>
 						<th>Name</th>
@@ -246,45 +255,56 @@
 					</tr>
 				</thead>
 				<tbody>
-					{#each tasks as task}
-						<tr>
-							<td class="col-3">{task.name}</td>
-							<td class="col-1">{task.version || '–'}</td>
-							<td class="col-1">{task.owner || '–'}</td>
-							<td class="col-2">
-								<button
-									data-fc-task={task.id}
-									class="btn btn-light"
-									data-bs-toggle="modal"
-									data-bs-target="#taskInfoModal"
-									on:click={setTaskModal}
-								>
-									<i class="bi bi-info-circle" />
-									Info
-								</button>
-								<button
-									on:click={() => {}}
-									data-fc-task={task.id}
-									class="btn btn-primary"
-									data-bs-toggle="modal"
-									data-bs-target="#taskEditModal"
-									on:click={setTaskModal}
-								>
-									<i class="bi bi-pencil" />
-									Edit
-								</button>
-								<ConfirmActionButton
-									modalId="confirmTaskDeleteModal{task.id}"
-									style={'danger'}
-									btnStyle="danger"
-									buttonIcon="trash"
-									label={'Delete'}
-									message={`Delete task ${task.name}`}
-									callbackAction={() => handleDeleteTask(task.id)}
-								/>
-							</td>
-						</tr>
-					{/each}
+					{#key tasks}
+						{#each tasks as task, i}
+							<tr
+								class:old-version={isOldVersion(i)}
+								class:last-old-version={isLastOldVersion(i)}
+								class:is-main-version={isMainVersion(i)}
+								class:collapsed={isOldVersion(i)}
+							>
+								<td class="col-3">{isOldVersion(i) ? '' : task.name}</td>
+								<td class="col-1">
+									{task.version || '–'}
+									{#if isMainVersion(i)}
+										<button class="btn btn-link" on:click={handleToggleOldVersions}>
+											<i class="bi bi-plus-circle" />
+										</button>
+									{/if}
+								</td>
+								<td class="col-1">{task.owner || '–'}</td>
+								<td class="col-2">
+									<button
+										class="btn btn-light"
+										data-bs-toggle="modal"
+										data-bs-target="#taskInfoModal"
+										on:click={() => setTaskModal(task.id)}
+									>
+										<i class="bi bi-info-circle" />
+										Info
+									</button>
+									<button
+										class="btn btn-primary"
+										data-bs-toggle="modal"
+										data-bs-target="#taskEditModal"
+										on:click={() => setTaskModal(task.id)}
+									>
+										<i class="bi bi-pencil" />
+										Edit
+									</button>
+									<ConfirmActionButton
+										modalId="confirmTaskDeleteModal{task.id}"
+										style={'danger'}
+										btnStyle="danger"
+										buttonIcon="trash"
+										label={'Delete'}
+										message={`Delete task ${task.name}`}
+										callbackAction={() => handleDeleteTask(task.id)}
+									/>
+								</td>
+							</tr>
+						{/each}
+					{/key}
 				</tbody>
 			</table>
 		</div>
@@ -293,3 +313,25 @@
 
 <TaskInfoModal />
 <TaskEditModal {updateEditedTask} />
+
+<style>
+	:global(.is-main-version.expanded td) {
+		border-bottom-style: dashed;
+	}
+
+	.old-version.collapsed {
+		display: none;
+	}
+
+	.old-version {
+		display: table-row;
+	}
+
+	.old-version td {
+		border-bottom-style: dashed;
+	}
+
+	.last-old-version td {
+		border-bottom-style: solid;
+	}
+</style>